code/training/requirements.txt
transformers
trl
peft
accelerate
bitsandbytes
huggingface-hub
jsonlines
pandas
protobuf==3.20.1
sentencepiece==0.2.0
tensorboard
torch
vllm==v0.4.3
vllm-flash-attn
nltk
fraction
PyYAML
typing_extensions
openai==0.28
autoawq
auto-gptq
